requirements.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
Pillow
psycopg2-binary
PyJWT
python-decouple
gunicorn
django-cors-headers